// apps/data-engine/src/models/marketitem-string.model.ts
export interface MarketItemString {
  timestamp: string;
  openPrice: number;
  highPrice: number;
  lowPrice: number;
  closePrice: number;
  volume: string;
  sma20: string | undefined;
  sma50: string | undefined;
  ema20: string | undefined;
  ema50: string | undefined;
  ema100: string | undefined;
  ema200: string | undefined;
  bbands20: string | undefined;
  rsi14: string | undefined;
  atr14: string | undefined;
  macdSignal: string | undefined;
  macdFast: string | undefined;
  macdSlow: string | undefined;
  macdNorm: string | undefined;
  slopeEma: string | undefined;
  distanceEma: string | undefined;
  roc: string | undefined;
  stoch: string | undefined;
  obv: string | undefined;
  volumeZScore: string | undefined;
  volumeAtrRatio: string | undefined;
  fvgBullish: string | undefined;
  fvgBearish: string | undefined;
  fvgSize: string | undefined;
  fvgSizeAtrNorm: string | undefined;
  bodySizePerc: string | undefined;
  upperWickPerc: string | undefined;
  lowerWickPerc: string | undefined;
  rangeExp: string | undefined;
  logReturn: string | undefined;
  rollingVolatility: string | undefined;
  rollingVolatilityAtrNorm: string | undefined;
  rollingVolatilitySlope: string | undefined;
}